-moved EventsModel interface into useEvents.tsx
